refactor: circular dependancy 이슈 관련 수정
