+++ src/js/DOM.js
@@ -1,0 +1,1 @@
+export const totalDisplayText = document.getElementById('total');
